Add publishEvent(s) to EventBus interface
Add publishEvent() and publishEvents() to the eventBus interface for
cases where the events are not being generated within an event
handler, but are instead the result of some external event that
did not come from the event bus (user action, external system
event).

Also refactor event bus code for readability and ease of use.

Add more event bus unit tests to ensure everything is working
as expected.